requirements.txt
# ====================================================================
# LSTM Battery Analyzer - 项目依赖包
# ====================================================================
# 
# 这个文件包含了整个项目运行所需的Python依赖包
# 建议在虚拟环境中安装这些依赖
#
# 安装方法：
# pip install -r requirements.txt
#
# ====================================================================

# ============= 核心Web框架 =============
fastapi>=0.100.0                # 现代异步Web框架
uvicorn[standard]>=0.22.0       # ASGI服务器
python-socketio>=5.8.0          # WebSocket实时通信
websockets>=11.0                # WebSocket协议支持
starlette>=0.27.0               # ASGI工具包

# ============= 数据处理和科学计算 =============
numpy>=1.24.0                   # 数值计算基础库
pandas>=2.0.0                   # 数据分析和处理
scikit-learn>=1.3.0             # 机器学习工具包（预处理、指标等）

# ============= 深度学习框架 =============
tensorflow>=2.13.0              # 深度学习框架（CNN+LSTM模型）
# 注意：如果您的系统支持GPU加速，可以安装 tensorflow-gpu

# ============= 数据可视化 =============
matplotlib>=3.7.0               # 绘图库（用于训练结果可视化）

# ============= HTTP和网络 =============
requests>=2.31.0                # HTTP客户端（测试和外部API调用）
aiohttp>=3.8.0                  # 异步HTTP客户端
python-multipart>=0.0.6         # 文件上传支持

# ============= 数据验证和序列化 =============
pydantic>=2.0.0                 # 数据验证和设置管理

# ============= 文件处理 =============
# zipfile-deflate64>=0.2.0      # 高级ZIP文件处理（可选）

# ============= 开发和测试工具 =============
# pytest>=7.4.0                 # 测试框架（开发时可选）
# pytest-asyncio>=0.21.0        # 异步测试支持

# ============= 性能优化（可选）=============
# 以下包可以提升性能，但不是必需的：
# psutil>=5.9.0                 # 系统资源监控
# numba>=0.57.0                 # 数值计算JIT编译加速
# joblib>=1.3.0                 # 并行计算加速

# ============= 系统依赖说明 =============
# 1. Python版本：推荐 Python 3.8+ 
# 2. 操作系统：支持 Windows、Linux、macOS
# 3. 内存要求：推荐 8GB+ RAM（用于深度学习模型训练）
# 4. 存储空间：推荐 2GB+ 可用空间（用于模型和数据存储）
#
# GPU支持（可选）：
# - 如需GPU加速，请安装 CUDA 和 cuDNN
# - TensorFlow会自动检测并使用可用的GPU